+++ src/features/product/productList/components/filters/FilterAttribute.tsx
@@ -1,8 +1,17 @@
-import useFilterAttribute from '../../hooks/filters/useFilterAttribute';
-import WidgetTitle from '@components/Widgets/Title';
 import Checkbox from '@components/Form/Checkbox';
+import WidgetTitle from '@components/Widgets/Title';
+import type { ActiveFilters, FilterCounts } from '@typings/products/filter';
 
-export default function FilterAttribute({ attributeType, activeFilters, filterCounts, handleFilterListToggle }) {
+import useFilterAttribute from '../../hooks/filters/useFilterAttribute';
+
+interface FilterAttributeProps {
+  attributeType: 'colour' | 'size';
+  activeFilters: ActiveFilters;
+  filterCounts: FilterCounts;
+  handleFilterListToggle: (filterKey: string, value: number) => void;
+}
+
+export default function FilterAttribute({ attributeType, activeFilters, filterCounts, handleFilterListToggle }: FilterAttributeProps) {
   const filterAttributes = useFilterAttribute(attributeType);
 
   return (
@@ -19,8 +28,8 @@ export default function FilterAttribute({ attributeType, activeFilters, filterCo
                 label={
                   <div className="flex items-center gap-x-1">
                     <span>{attribute.name}</span>
-                    {attributeType === 'colour' && <span>({filterCounts.colour[attribute.id] || 0})</span>}
-                    {attributeType === 'size' && <span>({filterCounts.size[attribute.id] || 0})</span>}
+                    {filterCounts && attributeType === 'colour' && <span>({filterCounts.colour[attribute.id] || 0})</span>}
+                    {filterCounts && attributeType === 'size' && <span>({filterCounts.size[attribute.id] || 0})</span>}
                   </div>
                 }
                 checked={activeFilters[attributeType].includes(attribute.id)}